+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Stricter option validation using `nimble_options`.
+- `sync_all` requires `types` option.
 - Rename `sanity_config` option to `request_opts` for consistency with `Sanity.stream/1`.
 
 ## [0.2.1] - 2022-10-23